Use bullseye for golang image, and set node at 10.24.1
Bullseye gives us Python 3.9, bookworm is the default for golang:1.19.10, but bookworm has Python 3.11, and in Python 3.10, a 'U' mode was removed which breaks node-gyp, which is required for the netlinkwrapper to work.  Netlinkwrapper is required to be able to reliably open a ws connection to cerveau.  The fallback syncsocket was not reliable, and failed to connect most times.

Node 10.24.1 was the latest version which was able to install netlinkwrapper@1.1.2

See https://github.com/nodejs/node-gyp/issues/2219